Move out tonic `UnixStream` wrapper into utils
